+++ etc/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+### Added
+- Enum StressDirection that indicates if the stress is increasing or decreasing.
+- Property StressDirection in AnalyzeResultsRequest.
+### Changed
+- Operation AnalyzeResults to also consider when the stress increase.
+
 ## [2.1.2] - 2021-04-26
 ### Fixed
 - Operation AnalyzeResults to not write the first and second derivative.